crop_disease_detection_app/requirements.txt
Django==3.2.10
djangorestframework==3.12.4
requests==2.26.0
wheel==0.33.6
pillow==9.0.0
django-cors-headers==3.10.1